Add a generic model-creating step
This accepts a count and a singular or plural factory name and creates
the requested number of records when called.  I'm adding this here and
now in a separate commit as I find that *everyone* seems to duplicate
this type of step multiple times when adding their own features.  I'm
hoping that having this here will prevent that.